Moved all code pertaining to sfr_models subpackage
