+++ halotools/empirical_models/__init__.py
@@ -9,9 +9,10 @@ from .preloaded_models import *
 from .hod_components import *
 from .model_defaults import *
 from .smhm_components import *
-from .sfr_components import *
 from .abunmatch import *
 from .model_helpers import *
 from .assembias_models import *
 from .hod_designer import *
+
 from .phase_space_marf import *
+from .sfr_models import *
