+++ halotools/empirical_models/sfr_models/__init__.py
@@ -1,0 +1,1 @@
+from .sfr_components import *
